+++ crates/wdl-ast/src/v1/validation.rs
@@ -2,10 +2,12 @@
 
 mod counts;
 mod keys;
+mod names;
 mod numbers;
 mod strings;
 
 pub use counts::*;
 pub use keys::*;
+pub use names::*;
 pub use numbers::*;
 pub use strings::*;